fix(reconciliation): move expandReconcileWindow out of "use server" file (#543)
Next.js 16 rejects non-async exports from a "use server" module — the
build fails with "Server Actions must be async functions". Splitting
the helper into a sibling window.ts keeps actions.ts pure server
actions and makes the helper independently importable from tests.
